deps: V8: cherry-pick d83f479604c8
Original commit message:

    Fix compilation for older version of libstdc++

    On older versions of libstdc++, cmath didn't expose std::powl.
    Use std::pow instead.

    Refs: https://gcc.gnu.org/pipermail/libstdc++/2023-February/055493.html
    Refs: https://github.com/nodejs/node/pull/61898
    Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7666244
    Commit-Queue: Joyee Cheung <joyee@igalia.com>
    Cr-Commit-Position: refs/heads/main@{#105884}

Refs: https://github.com/v8/v8/commit/d83f479604c8cf0e74b7e73e795adbd6f86549f2
PR-URL: https://github.com/nodejs/node/pull/61898

+++ deps/v8/src/objects/js-duration-format.cc
@@ -807,7 +807,7 @@ void OutputFractional(const char* type, int64_t integer, int32_t powerOfTen,
   // Pass in the value as int64_t and ask ICU to scale down.
   nfOpts = nfOpts.scale(icu::number::Scale::powerOfTen(-powerOfTen));
 
-  int64_t factor = static_cast<int64_t>(std::powl(10, powerOfTen));
+  int64_t factor = static_cast<int64_t>(std::pow(10.0L, powerOfTen));
   int64_t bound = std::numeric_limits<int64_t>::max() / factor - 1;
   UErrorCode status = U_ZERO_ERROR;
   // Use faster ICU API formatInt if the value fit the precision int64_t,
